incapsulate table init
